refactor(studyguides): split 974-line service.go into list/read/write files
Pure file-shuffle, zero behavior change. service.go grew from ~660
lines (PR #137) to 974 (after PR #140) across the studyguides write
surface buildout. Was flagged in PR #138 (L3), #139 (L2), and
#140 (M1). Splitting now while the seams are clean.

New file layout (all in package studyguides):
  service.go       106  -- Repository interface, Service struct,
                          NewService constructor + queryTable wiring
  service_list.go  346  -- ListStudyGuides, dbFilters, the 8
                          per-sort-variant query methods, cursor
                          helpers (cursorInt64/Timestamp/ID),
                          buildCursor, mapListRows, escapeLikePattern,
                          toDBFilters, validateListParams
  service_read.go  162  -- AssertCourseExists, GetStudyGuide
                          (with the 5-way errgroup fan-out)
  service_write.go 408  -- CreateStudyGuide, DeleteStudyGuide,
                          CastVote, RemoveVote, RecommendStudyGuide,
                          RemoveRecommendation, plus write-side
                          helpers (normalizeTags, trimmedNonEmpty,
                          validateCreateParams, guideVoteToDB)

Tests stay in service_test.go; mocks unaffected. go build, go vet,
go test -race all green with no test changes.
